+++ src/components/pages/MainPageCard/index.tsx
